attendance_system/deployment/requirements_production.txt
# Core Framework
flask==3.1.2
werkzeug==3.1.4

# Database
mongoengine
pymongo

# Authentication & Forms
flask-login
flask-wtf
wtforms
email_validator

# Machine Learning & Computer Vision
opencv-contrib-python>=4.11.0
numpy>=2.3.4
pillow

# Production Server
gunicorn==21.2.0

# Environment Management
python-dotenv==1.0.0

# Security (Optional but Recommended)
flask-talisman==1.1.0  # HTTPS enforcement and security headers

# Monitoring & Logging (Optional)
# sentry-sdk[flask]==1.40.0  # Error tracking
# flask-limiter==3.5.0  # Rate limiting

# Caching (Optional)
# flask-caching==2.1.0
# redis==5.0.1

# Background Tasks (Optional)
# celery==5.3.4
# redis==5.0.1

# Testing (Development)
# pytest==7.4.3
# pytest-flask==1.3.0
# pytest-cov==4.1.0
